Reapply c9965ac
